Fix for layout regressions from glyph font consolidation.
Made all glyphs in the font rest on the baseline and be
properly horizonally centered.

Previously, there were baseline offsets in the font *and*
in the app code. Now offsets are controlled by the app code
only.

Removed the font/svg scripts as they were out of date and
are maintained at:
https://github.com/montehurd/FontToSvgsToFont

Change-Id: Ibd8f78f6b30858889483958d6ed0cf859a1746e4