WIP: Native component beneath webView.
Discovered better way to make native view track web view
with autolayout. Factored to category on UIWebView for
easy top or bottom placement of tracking view.

Made lead image use the new category.

Added bottom container as well with stubs for
3 sub-containers and their respective controllers.

The top-most of the sub-containers is a stub
for "suggestions".

The other two will be finished as part of this story.

Change-Id: I60f1ac5b9ffbe5eb1e75c0d30926063a884019c0