Move lead image inflation/face detection to background thread.
Listens in on URLCache notification that the web view retrieved
an image and sees if this image is a variant of the lead image.
Shows this image if the variant is larger than the one already
being shown.

Fixed performance issue in data layer related to seeing if an
image binary exists.

Change-Id: I979f3a07c7d37febd5b7d9f99a8105223f2525b3