readme: update to be more user-friendly
Lots of project dependencies were removed by prior commits. These
changes reflect the new, streamlined workflow and the fact that devs
should now be able to simply clone and run the project.

Also, broke out lots of information to separate files in the "docs"
folder to make the README leaner.

Change-Id: I33df4b6b2917063e6e78b228c2f29e901eb153fa